+++ truck-rendimpl/tests/common.rs
@@ -177,7 +177,7 @@ pub fn init_device(instance: &Instance) -> DeviceHandler {
             })
             .await
             .unwrap();
-        DeviceHandler::new(Arc::new(adapter), Arc::new(device), Arc::new(queue))
+        DeviceHandler::new(adapter, device, queue)
     })
 }